Extract out generic unit tests from cuda/test_tc_mapper.cc
These will be reused with the upcoming CPU TC mapper.
