+++ test/cuda/CMakeLists.txt
@@ -1,4 +1,5 @@
 include_directories(.)
+include_directories(..)
 
 set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
